01_etl/requirements.txt
# python 3.9.7

elasticsearch==7.17.3
flake8==4.0.1
psycopg2-binary==2.9.3
pydantic==1.9.0
python-dotenv==0.20.0
redis==4.2.2
